prevent restart for rclone if disabled

+++ run/nobody/logrotate.sh
@@ -8,41 +8,49 @@ log_path="${1}"
 number_of_logs_to_keep=3
 file_size_limit_kb=10240
 
-# wait for log file to exist before size checks proceed
-while [[ ! -f "${log_path}" ]]; do
-	sleep 0.1
-done
+if [[ "${ENABLE_RCLONE}" == 'no' ]]; then
 
-while true; do
+	echo "[info] rclone not enabled, disabled log rotate"
 
-	file_size_kb=$(du -k "${log_path}" | cut -f1)
+else
 
-	if [ "${file_size_kb}" -ge "${file_size_limit_kb}" ]; then
+	# wait for log file to exist before size checks proceed
+	while [[ ! -f "${log_path}" ]]; do
+		sleep 0.1
+	done
 
-		echo "[info] '${log_path}' log file larger than limit ${file_size_limit_kb} kb, rotating logs..."
+	while true; do
 
-		if [[ -f "${log_path}.${number_of_logs_to_keep}" ]]; then
-			echo "[info] Deleting oldest log file '${log_path}.${number_of_logs_to_keep}'..."
-			rm -f "${log_path}.${number_of_logs_to_keep}"
-		fi
+		file_size_kb=$(du -k "${log_path}" | cut -f1)
 
-		for log_number in $(seq "${number_of_logs_to_keep}" -1 0); do
+		if [ "${file_size_kb}" -ge "${file_size_limit_kb}" ]; then
 
-			if [[ -f "${log_path}.${log_number}" ]]; then
-				log_number_inc=$((log_number+1))
-				mv "${log_path}.${log_number}" "${log_path}.${log_number_inc}"
+			echo "[info] '${log_path}' log file larger than limit ${file_size_limit_kb} kb, rotating logs..."
+
+			if [[ -f "${log_path}.${number_of_logs_to_keep}" ]]; then
+				echo "[info] Deleting oldest log file '${log_path}.${number_of_logs_to_keep}'..."
+				rm -f "${log_path}.${number_of_logs_to_keep}"
 			fi
 
-		done
+			for log_number in $(seq "${number_of_logs_to_keep}" -1 0); do
 
-		echo "[info] Copying current log '${log_path}' to ${log_path}.0..."
-		cp "${log_path}" "${log_path}.0"
+				if [[ -f "${log_path}.${log_number}" ]]; then
+					log_number_inc=$((log_number+1))
+					mv "${log_path}.${log_number}" "${log_path}.${log_number_inc}"
+				fi
 
-		echo "[info] Emptying current log '${log_path}' contents..."
-		> "${log_path}"
+			done
 
-	fi
+			echo "[info] Copying current log '${log_path}' to ${log_path}.0..."
+			cp "${log_path}" "${log_path}.0"
 
-	sleep 30s
+			echo "[info] Emptying current log '${log_path}' contents..."
+			> "${log_path}"
 
-done
+		fi
+
+		sleep 30s
+
+	done
+
+fi
